MSI-1029: Configurable Product Presents on Category Page When All It's Children Set to "Out of Stock"
