+++ app/code/Magento/InventoryCatalog/Plugin/InventoryApi/SetDataToLegacyCatalogInventoryAtSourceItemsSavePlugin.php
@@ -7,27 +7,15 @@ declare(strict_types=1);
 
 namespace Magento\InventoryCatalog\Plugin\InventoryApi;
 
-use Magento\CatalogInventory\Api\Data\StockItemInterface;
-use Magento\CatalogInventory\Api\StockItemRepositoryInterface;
-use Magento\CatalogInventory\Api\StockItemCriteriaInterfaceFactory;
-use Magento\CatalogInventory\Model\Indexer\Stock\Processor;
-use Magento\CatalogInventory\Model\Spi\StockStateProviderInterface;
-use Magento\CatalogInventory\Model\Stock;
-use Magento\Framework\Exception\InputException;
 use Magento\InventoryApi\Api\Data\SourceItemInterface;
 use Magento\InventoryApi\Api\SourceItemsSaveInterface;
 use Magento\InventoryCatalogApi\Api\DefaultSourceProviderInterface;
-use Magento\InventoryCatalogApi\Model\GetProductIdsBySkusInterface;
 use Magento\InventoryCatalogApi\Model\GetProductTypesBySkusInterface;
-use Magento\InventoryCatalog\Model\ResourceModel\SetDataToLegacyStockItem;
+use Magento\InventoryCatalogApi\Model\SourceItemsSaveSynchronizationInterface;
 use Magento\InventoryConfigurationApi\Model\IsSourceItemManagementAllowedForProductTypeInterface;
 
 /**
- * Set Qty and status for legacy CatalogInventory Stock Status and Stock Item DB tables,
- * if corresponding MSI SourceItem assigned to Default Source has been saved
- *
- * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
- * TODO: https://github.com/magento-engcom/msi/pull/1082/
+ * Synchronization between legacy Stock Item with Source Item
  */
 class SetDataToLegacyCatalogInventoryAtSourceItemsSavePlugin
 {
@@ -35,36 +23,6 @@ class SetDataToLegacyCatalogInventoryAtSourceItemsSavePlugin
      * @var DefaultSourceProviderInterface
      */
     private $defaultSourceProvider;
-
-    /**
-     * @var SetDataToLegacyStockItem
-     */
-    private $setDataToLegacyStockItem;
-
-    /**
-     * @var StockItemCriteriaInterfaceFactory
-     */
-    private $legacyStockItemCriteriaFactory;
-
-    /**
-     * @var StockItemRepositoryInterface
-     */
-    private $legacyStockItemRepository;
-
-    /**
-     * @var GetProductIdsBySkusInterface
-     */
-    private $getProductIdsBySkus;
-
-    /**
-     * @var StockStateProviderInterface
-     */
-    private $stockStateProvider;
-
-    /**
-     * @var Processor
-     */
-    private $indexerProcessor;
 
     /**
      * @var IsSourceItemManagementAllowedForProductTypeInterface
@@ -77,36 +35,26 @@ class SetDataToLegacyCatalogInventoryAtSourceItemsSavePlugin
     private $getProductTypeBySku;
 
     /**
+     * @var SourceItemsSaveSynchronizationInterface
+     */
+    private $sourceItemSynchronization;
+
+    /**
      * @param DefaultSourceProviderInterface $defaultSourceProvider
-     * @param SetDataToLegacyStockItem $setDataToLegacyStockItem
-     * @param StockItemCriteriaInterfaceFactory $legacyStockItemCriteriaFactory
-     * @param StockItemRepositoryInterface $legacyStockItemRepository
-     * @param GetProductIdsBySkusInterface $getProductIdsBySkus
-     * @param StockStateProviderInterface $stockStateProvider
-     * @param Processor $indexerProcessor
      * @param IsSourceItemManagementAllowedForProductTypeInterface $isSourceItemsAllowedForProductType
      * @param GetProductTypesBySkusInterface $getProductTypeBySku
+     * @param SourceItemsSaveSynchronizationInterface $sourceItemsSaveSynchronization
      */
     public function __construct(
         DefaultSourceProviderInterface $defaultSourceProvider,
-        SetDataToLegacyStockItem $setDataToLegacyStockItem,
-        StockItemCriteriaInterfaceFactory $legacyStockItemCriteriaFactory,
-        StockItemRepositoryInterface $legacyStockItemRepository,
-        GetProductIdsBySkusInterface $getProductIdsBySkus,
-        StockStateProviderInterface $stockStateProvider,
-        Processor $indexerProcessor,
         IsSourceItemManagementAllowedForProductTypeInterface $isSourceItemsAllowedForProductType,
-        GetProductTypesBySkusInterface $getProductTypeBySku
+        GetProductTypesBySkusInterface $getProductTypeBySku,
+        SourceItemsSaveSynchronizationInterface $sourceItemsSaveSynchronization
     ) {
         $this->defaultSourceProvider = $defaultSourceProvider;
-        $this->setDataToLegacyStockItem = $setDataToLegacyStockItem;
-        $this->legacyStockItemCriteriaFactory = $legacyStockItemCriteriaFactory;
-        $this->legacyStockItemRepository = $legacyStockItemRepository;
-        $this->getProductIdsBySkus = $getProductIdsBySkus;
-        $this->stockStateProvider = $stockStateProvider;
-        $this->indexerProcessor = $indexerProcessor;
         $this->isSourceItemsAllowedForProductType = $isSourceItemsAllowedForProductType;
         $this->getProductTypeBySku = $getProductTypeBySku;
+        $this->sourceItemSynchronization = $sourceItemsSaveSynchronization;
     }
 
     /**
@@ -118,73 +66,22 @@ class SetDataToLegacyCatalogInventoryAtSourceItemsSavePlugin
      */
     public function afterExecute(SourceItemsSaveInterface $subject, $result, array $sourceItems): void
     {
-        $productIds = [];
+        $sourceItemsForSynchronization = [];
         foreach ($sourceItems as $sourceItem) {
             if ($sourceItem->getSourceCode() !== $this->defaultSourceProvider->getCode()) {
                 continue;
             }
 
             $sku = $sourceItem->getSku();
-
-            try {
-                $productId = (int)$this->getProductIdsBySkus->execute([$sku])[$sku];
-            } catch (InputException $e) {
-                // Saving source item data for not existed product
-                continue;
-            }
-
             $typeId = $this->getProductTypeBySku->execute([$sku])[$sku];
+
             if (false === $this->isSourceItemsAllowedForProductType->execute($typeId)) {
                 continue;
             }
 
-            $legacyStockItem = $this->getLegacyStockItem($productId);
-            if (null === $legacyStockItem) {
-                continue;
-            }
-
-            $isInStock = (int)$sourceItem->getStatus();
-
-            if ($legacyStockItem->getManageStock()) {
-                $legacyStockItem->setIsInStock($isInStock);
-                $legacyStockItem->setQty((float)$sourceItem->getQuantity());
-
-                if (false === $this->stockStateProvider->verifyStock($legacyStockItem)) {
-                    $isInStock = 0;
-                }
-            }
-
-            $this->setDataToLegacyStockItem->execute(
-                $sourceItem->getSku(),
-                (float)$sourceItem->getQuantity(),
-                $isInStock
-            );
-            $productIds[] = $productId;
+            $sourceItemsForSynchronization[] = $sourceItem;
         }
 
-        if ($productIds) {
-            $this->indexerProcessor->reindexList($productIds);
-        }
-    }
-
-    /**
-     * @param int $productId
-     * @return null|StockItemInterface
-     */
-    private function getLegacyStockItem(int $productId): ?StockItemInterface
-    {
-        $searchCriteria = $this->legacyStockItemCriteriaFactory->create();
-
-        $searchCriteria->addFilter(StockItemInterface::PRODUCT_ID, StockItemInterface::PRODUCT_ID, $productId);
-        $searchCriteria->addFilter(StockItemInterface::STOCK_ID, StockItemInterface::STOCK_ID, Stock::DEFAULT_STOCK_ID);
-
-        $stockItemCollection = $this->legacyStockItemRepository->getList($searchCriteria);
-        if ($stockItemCollection->getTotalCount() === 0) {
-            return null;
-        }
-
-        $stockItems = $stockItemCollection->getItems();
-        $stockItem = reset($stockItems);
-        return $stockItem;
+        $this->sourceItemSynchronization->execute($sourceItemsForSynchronization);
     }
 }
